Move images to public folder
