+++ src/pages/getinvolved.js
@@ -5,19 +5,19 @@ import Button from 'react-bootstrap/Button'
 import Card from 'react-bootstrap/Card'
 import CardColumns from 'react-bootstrap/CardColumns'
 
-import logo3 from '../../images/ccollege.png'
-import logo5 from '../../images/crt.png'
-import logo4 from '../../images/CWR_logo.png'
-import logo6 from '../../images/deepgreen.png'
-import logo2 from '../../images/folks.png'
-import hackathon from '../../images/hackathon.png'
-import logo8 from '../../images/OI.png'
-import logo10 from '../../images/orcanetwork.png'
-import logo9 from '../../images/pacman.png'
-import roadmap from '../../images/roadmap.png'
-import orcas from '../../images/srkw2-10.jpg'
-import logo1 from '../../images/twt.png'
-import logo11 from '../../images/ws_logo.png'
+import logo3 from '../../public/images/ccollege.png'
+import logo5 from '../../public/images/crt.png'
+import logo4 from '../../public/images/CWR_logo.png'
+import logo6 from '../../public/images/deepgreen.png'
+import logo2 from '../../public/images/folks.png'
+import hackathon from '../../public/images/hackathon.png'
+import logo8 from '../../public/images/OI.png'
+import logo10 from '../../public/images/orcanetwork.png'
+import logo9 from '../../public/images/pacman.png'
+import roadmap from '../../public/images/roadmap.png'
+import orcas from '../../public/images/srkw2-10.jpg'
+import logo1 from '../../public/images/twt.png'
+import logo11 from '../../public/images/ws_logo.png'
 import getinvolvedStyles from '../styles/getinvolved.module.css'
 
 export const getinvolved = () => {
